+++ src/main/java/com/owczarczak/footballers/clubRepresentation/ClubRepresentationRepository.java
@@ -2,5 +2,5 @@ package com.owczarczak.footballers.clubRepresentation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClubRepresentationRepository extends JpaRepository<ClubRepresentation, Integer> {
+public interface ClubRepresentationRepository extends JpaRepository<ClubRepresentation, Long> {
 }
