+++ src/main/resources/db/migration/V2__added_clubs_table.sql
@@ -10,7 +10,7 @@ drop column if exists club,
 drop column if exists goals;
 
 create table club(
-	id serial,
+	id bigserial,
 	name varchar(255) not null,
 	created date,
 	primary key (id),
@@ -18,9 +18,9 @@ create table club(
 );
 
 create table contract(
-	id serial,
-	club_id int,
-	footballer_id int,
+	id bigserial,
+	club_id bigint,
+	footballer_id bigint,
 	contract_start date,
 	contract_end date,
 	salary int,
@@ -30,16 +30,16 @@ create table contract(
 );
 
 create table club_representation(
-	id serial ,
-	club_id int not null,
+	id bigserial ,
+	club_id bigint not null,
     primary key (id),
 	constraint fk_club foreign key(club_id) references club(id)
 );
 
 create table match(
-	id serial,
-	host_representation_id int,
-	guest_representation_id int,
+	id bigserial,
+	host_representation_id bigint,
+	guest_representation_id bigint,
 	name_of_referee varchar(255),
 	date timestamp,
     primary key (id),
@@ -50,17 +50,17 @@ create table match(
 );
 
 create table club_representation_footballer(
-	club_representation_id int not null,
-	footballer_id int not null,
+	club_representation_id bigint not null,
+	footballer_id bigint not null,
 	constraint fk_representation foreign key(club_representation_id) references club_representation(id),
 	constraint fk_footballer foreign key(footballer_id) references footballer(id),
 	unique(club_representation_id, footballer_id)
 );
 
 create table score(
-    id serial,
-	match_id int not null,
-	footballer_id int not null,
+    id bigserial,
+	match_id bigint not null,
+	footballer_id bigint not null,
 	minute_scored int,
 	constraint fk_match foreign key(match_id) references match(id),
 	constraint fk_footballer foreign key(footballer_id) references footballer(id)
